Kosh (#257)
* ♻️ Refactor: 타임라인 바 위치 분배 로직 util 함수로 모듈화

* 🔧 Fix: 타임라인 API 카테고리 변경에 따른 수정

* 🔧 Fix: 모듈화를 통해 내커리어 페이지 색상 통일화
